feat: 🎸 [FrontEnd] optimize code
